Moved templates to content dir
